requirements.txt
Django == 4.2
django-environ==0.10.0
django-stubs-ext == 0.8.0
django-cleanup == 7.0.0
django-stubs == 1.16.0
django-debug-toolbar == 4.0.0
djlint == 1.23.3
django-crispy-forms == 2.0
django_countries == 7.5.1
django-colorfield == 0.8.0
django-phonenumber-field[phonenumbers] == 7.1.0
django-simple-captcha == 0.5.17 
crispy-bootstrap5 == 0.7
psycopg2-binary == 2.9.6
Pillow == 9.5.0
black == 23.3.0
isort == 5.12.0
mypy == 1.2.0